Complete GPU based alpha premultiplication impl and make the CPU version even faster.
* GPU based version started in previous commit, but this fixes errors
  and bugs and gets it actually compiling and running.
* Add a way to batch together images to use the same render pass for GPU
  premultiplication if they all target the same texture.
* Pending premultiplication uploads are automatically done when calling
  `Drawer::third_pass`.
* `fast-srgb8` dep removed, we no longer convert to `f32`s to do the
  premultiplication. Two `[u16; 256]` tables are combined to compute the
  alpa premultiplied color within the same error bounds used by the
  `fast-srgb8` crate. We also no longer use explicit simd.
* Remove explicit lifetimes from `PlayState::render` since `&self` and
  `Drawer<'_>` don't need to have the same lifetime.
* Fix existing bug where invalidated cache entries were never set to
  valid when reusing them.
* `prepare_graphic` now runs some heuristics to determine whether
  premultiplication should be executed CPU side or GPU side and then
  returns a bool indicating if GPU premultiplication is needed.

+++ assets/voxygen/shaders/include/srgb.glsl
@@ -44,7 +44,7 @@ vec3 linear_to_srgb(vec3 col) {
 }
 
 vec4 srgba8_to_linear(uint srgba8) {
-    uvec4 nonlinear = vec4(uvec4(
+    vec4 nonlinear = vec4(uvec4(
         (srgba8 >> 24) & 0xFFu,
         (srgba8 >> 16) & 0xFFu,
         (srgba8 >>  8) & 0xFFu,
